create separate spec for stateRenderer-v2
